requirements.txt
affine==2.4.0
alembic==1.15.2
amqp==5.3.1
attrs==25.3.0
billiard==4.2.1
blinker==1.9.0
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.5
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
GeoAlchemy2==0.17.1
geopandas==1.0.1
greenlet==3.2.0
gunicorn==25.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.5.3
Mako==1.3.10
mapclassify==2.10.0
MarkupSafe==3.0.2
networkx==3.6.1
numpy==2.2.5
packaging==25.0
pandas==2.2.3
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyogrio==0.10.0
pyparsing==3.2.3
pyproj==3.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
rasterio==1.4.3
redis==6.0.0
requests==2.32.5
scikit-learn==1.8.0
scipy==1.15.2
shapely==2.1.0
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.6.3
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
